Remove old demo data source page
