+++ packages/front-end/components/DemoDataSourceGlobalBanner/DemoDataSourceGlobalBanner.tsx
@@ -1,14 +1,30 @@
 import React, { FC, useCallback, useMemo, useState } from "react";
+import { Flex } from "@radix-ui/themes";
 import { useRouter } from "next/router";
 import { PiCaretDownFill } from "react-icons/pi";
 import { getDemoDatasourceProjectIdForOrganization } from "shared/demo-datasource";
-import { deleteDemoDatasource } from "@/components/DemoDataSourcePage/DemoDataSourcePage";
 import { useDefinitions } from "@/services/DefinitionsContext";
-import { useAuth } from "@/services/auth";
+import { AuthContextValue, useAuth } from "@/services/auth";
 import track from "@/services/track";
 import { Popover } from "@/ui/Popover";
 import Button from "@/ui/Button";
 import Callout from "@/ui/Callout";
+
+export async function deleteDemoDatasource(
+  apiCall: AuthContextValue["apiCall"],
+) {
+  await apiCall(`/demo-datasource-project`, {
+    method: "DELETE",
+  });
+}
+
+export async function resetDemoDatasource(
+  apiCall: AuthContextValue["apiCall"],
+) {
+  await apiCall(`/demo-datasource-project/reset`, {
+    method: "POST",
+  });
+}
 
 type DemoDataSourceGlobalBannerProps = {
   ready: boolean;
@@ -24,10 +40,12 @@ export const DemoDataSourceGlobalBanner: FC<
   const { mutateDefinitions, project, projects, setProject } = useDefinitions();
   const router = useRouter();
   const [error, setError] = useState<string | null>(null);
+  const [success, setSuccess] = useState<string | null>(null);
   const [open, setOpen] = useState(false);
 
   const handleDelete = useCallback(async () => {
     setError(null);
+    setSuccess(null);
     try {
       track("Delete Sample Project", { source: "global-banner" });
       await deleteDemoDatasource(apiCall);
@@ -58,6 +76,19 @@ export const DemoDataSourceGlobalBanner: FC<
     setProject,
   ]);
 
+  const handleReset = useCallback(async () => {
+    setError(null);
+    setSuccess(null);
+    try {
+      track("Reset Sample Project", { source: "global-banner" });
+      await resetDemoDatasource(apiCall);
+      mutateDefinitions();
+      setSuccess("Sample data was reset to its original state.");
+    } catch (e) {
+      setError(e instanceof Error ? e.message : "Failed to reset sample data.");
+    }
+  }, [apiCall, mutateDefinitions]);
+
   if (!ready || !currentProjectIsDemo) {
     return null;
   }
@@ -83,16 +114,28 @@ export const DemoDataSourceGlobalBanner: FC<
               </button>
             }
             content={
-              <div style={{ maxWidth: 360 }}>
-                <p>If you are done with this sample data, delete it here.</p>
+              <div style={{ maxWidth: 380 }}>
+                <p>
+                  Mess something up? Reset it. Delete when you&apos;re done.
+                </p>
                 {error && (
                   <Callout status="error" mb="2">
                     {error}
                   </Callout>
                 )}
-                <Button color="red" onClick={handleDelete}>
-                  Delete Sample Data
-                </Button>
+                {success && (
+                  <Callout status="success" mb="2">
+                    {success}
+                  </Callout>
+                )}
+                <Flex gap="2">
+                  <Button variant="outline" onClick={handleReset}>
+                    Reset Sample Data
+                  </Button>
+                  <Button color="red" onClick={handleDelete}>
+                    Delete Sample Data
+                  </Button>
+                </Flex>
               </div>
             }
           />
