--- packages/front-end/pages/demo-datasource-project.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import { DemoDataSourcePageContainer } from "@/components/DemoDataSourcePage/DemoDataSourcePage";
-
-const DemoDataSourceProjectPage: NextPage = () => {
-  return <DemoDataSourcePageContainer />;
-};
-
-export default DemoDataSourceProjectPage;
